+++ src/main/kotlin/io/github/realyusufismail/armourandtoolsmod/core/init/ItemInit.kt
@@ -21,7 +21,9 @@ package io.github.realyusufismail.armourandtoolsmod.core.init
 import io.github.realyusufismail.armourandtoolsmod.MOD_ID
 import io.github.realyusufismail.armourandtoolsmod.core.armour.*
 import io.github.realyusufismail.armourandtoolsmod.core.material.ArmourMaterialInit
+import io.github.realyusufismail.armourandtoolsmod.core.material.CustomShieldMaterial
 import io.github.realyusufismail.armourandtoolsmod.core.material.CustomToolMaterial
+import io.github.realyusufismail.armourandtoolsmod.core.shields.*
 import net.minecraft.world.item.*
 import net.minecraftforge.registries.DeferredRegister
 import net.minecraftforge.registries.ForgeRegistries
@@ -182,8 +184,6 @@ object ItemInit {
             RainbowArmour(ArmourMaterialInit.RAINBOW, ArmorItem.Type.BOOTS, Item.Properties())
         }
 
-    // TODO: Add tools and weapons including tridents. Maybe custom shields?
-
     // Swords
     val RUBY_SWORD: ObjectHolderDelegate<Item> =
         ITEMS.registerSmeltableObject("ruby_sword", RUBY) {
@@ -234,27 +234,27 @@ object ItemInit {
     // Shields
     val RUBY_SHIELD: ObjectHolderDelegate<Item> =
         ITEMS.registerSmeltableObject("ruby_shield", RUBY) {
-            ShieldItem(Item.Properties().stacksTo(1).durability(600))
+            ArmourToolsModShieldItem(410, CustomShieldMaterial.RUBY_SHIELD)
         }
 
     val SAPPHIRE_SHIELD: ObjectHolderDelegate<Item> =
         ITEMS.registerSmeltableObject("sapphire_shield", SAPPHIRE) {
-            ShieldItem(Item.Properties().stacksTo(1).durability(600))
+            ArmourToolsModShieldItem(420, CustomShieldMaterial.SAPPHIRE_SHIELD)
         }
 
     val GRAPHITE_SHIELD: ObjectHolderDelegate<Item> =
         ITEMS.registerSmeltableObject("graphite_shield", GRAPHITE) {
-            ShieldItem(Item.Properties().stacksTo(1).durability(600))
+            ArmourToolsModShieldItem(400, CustomShieldMaterial.GRAPHITE_SHIELD)
         }
 
     val AQUMARINE_SHIELD: ObjectHolderDelegate<Item> =
         ITEMS.registerSmeltableObject("aqumarine_shield", AQUMARINE) {
-            ShieldItem(Item.Properties().stacksTo(1).durability(600))
+            ArmourToolsModShieldItem(450, CustomShieldMaterial.AQUMARINE_SHIELD)
         }
 
     val RAINBOW_SHIELD: ObjectHolderDelegate<Item> =
         ITEMS.registerSmeltableObject("rainbow_shield", RAINBOW) {
-            ShieldItem(Item.Properties().stacksTo(1).durability(600))
+            ArmourToolsModShieldItem(430, CustomShieldMaterial.RAINBOW_SHIELD)
         }
 
     private fun DeferredRegister<Item>.registerSmeltableObject(